+++ src/renderer/src/model/AppDataManager/DataClasses/SettingsData.ts
@@ -5,7 +5,7 @@ import { TxSettingsData } from './TxSettingsData';
 import { GeneralSettingsConfig } from './GeneralSettingsData';
 import { GraphingSettingsData } from './GraphingSettingsData';
 import { LogSettingsData } from './LogSettingsData';
-import { SoundsSettingsData } from './SoundsSettingsData';
+import { RulesSettingsData } from './RulesSettingsData';
 
 /**
  * Encapsulates all application settings data.
@@ -19,5 +19,5 @@ export class SettingsData {
   generalSettings = new GeneralSettingsConfig();
   graphingSettings = new GraphingSettingsData();
   logSettings = new LogSettingsData();
-  soundsSettings = new SoundsSettingsData();
+  rulesSettings = new RulesSettingsData();
 }