+++ src/renderer/src/model/Settings/Settings.tsx
@@ -8,7 +8,7 @@ import DisplaySettings from './DisplaySettings/DisplaySettings';
 import { PortSettings } from './PortSettings/PortSettings';
 import GeneralSettings from './GeneralSettings/GeneralSettings';
 import ProfilesSettings from './ProfileSettings/ProfileSettings';
-import SoundsSettings from './SoundsSettings/SoundsSettings';
+import RulesSettings from './RulesSettings/RulesSettings';
 import { App } from '../App';
 
 export enum SettingsCategories {
@@ -18,7 +18,7 @@ export enum SettingsCategories {
   DISPLAY,
   GENERAL,
   PROFILES,
-  SOUNDS,
+  RULES,
 }
 
 export class Settings {
@@ -40,7 +40,7 @@ export class Settings {
 
   profilesSettings: ProfilesSettings;
 
-  soundsSettings: SoundsSettings;
+  rulesSettings: RulesSettings;
 
   /**
    * Constructor for the Settings class.
@@ -57,7 +57,7 @@ export class Settings {
     this.displaySettings = new DisplaySettings(this.app.profileManager);
     this.generalSettings = new GeneralSettings(this.app.profileManager);
     this.profilesSettings = new ProfilesSettings(this.app.profileManager);
-    this.soundsSettings = new SoundsSettings(this.app.profileManager);
+    this.rulesSettings = new RulesSettings(this.app.profileManager);
     makeAutoObservable(this); // Make sure this is at the end of the constructor
   }